refactor: remove duplicate get_padding
